+++ src/main/java/com/gianca1994/aowebbackend/resources/npc/dto/request/NpcDTO.java
@@ -17,8 +17,8 @@ import lombok.Setter;
 public class NpcDTO {
     private String name;
     private short level;
-    private Long giveMaxExp;
-    private Long giveMinExp;
+    private int giveMaxExp;
+    private int giveMinExp;
     private Long giveMaxGold;
     private Long giveMinGold;
     private int hp;
